Remove create_admin from auth.py
